// sourcecode/BTrak.Packages/snova-billing-module/billing-module-components/billing-components/src/lib/billing/models/bl-model.ts
export class BLModel {
    bLNumber: string;
    purchaseExecutionId: string;
    bLDate: any;
    bLQuantity: any;
    chaId: string;
    consignerId: string;
    consigneeId: string;
    notifyParty: string;
    packingDetails: string;
    isDocumentsSent: boolean;
    sentDate: any;
    draftEntryDate: any;
    draftBLNumber: string;
    draftBLDescription: string;
    draftBasicCustomsDuty: string;
    draftSWC: any;
    draftIGST: any;
    draftEduCess: any;
    draftOthers: any;
    confoEntryDate: any;
    confoBLNumber: string;
    confoBLDescription: string;
    confoBasicCustomsDuty: string;
    confoSWC: any;
    confoIGST: any;
    confoEduCess: any;
    confoOthers: any;
    isConfirmedBill: string;
    initialDocumentsDescriptionsXml: string;
    finalDocumentsDescriptionsXml: string;
    initialDocumentsDescription: any[]
    finalDocumentsDescription: any[]
    confirmationDate: any;
    confoIsPaymentDone: boolean;
    confoPaymentDate: any;
	isArchived: boolean;
    shipmentBLId: string;
    blNumber: string;
    isInitialDocumentsMail: boolean;
}